requirements.txt
ultralytics==8.3.89
zenml==0.80.0
patool==4.0.0
opencv-python==4.11.0.86
numpy==2.1.1
tqdm==4.67.1
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.8.2
requests==2.32.3